phy: phy-rockchip-inno-usb2: correct the error of address for rv1108
The driver of u2phy use either grf or usbgrf to detect charger type.
But the register is divide into two parts both in grf and usbgrf in
rv1108. So rebase the regitster in usbgrf to grf register to make
sure that the driver can use a same base address.

Change-Id: I90ac8164b4dce2412a21ed7b5a5180dca1e9af3c
Signed-off-by: MengDongyang <daniel.meng@rock-chips.com>