spl: fit: use ddr for mmc read buffer in SPL
SPL need read data to IRAM for bl31(m0&data), but the SDHCI does
not support this path, we need to use DDR space as buffer.
The SDHCI driver do not have any error/warning if we set the
destination address to IRAM, and complete seems normally, but
the data is not transfer.

Change-Id: Ied9eed835961470bc87fda99a1565992c4103d9c
Signed-off-by: Kever Yang <kever.yang@rock-chips.com>